Extracted KeywordSeparator from GroupsPreferences and created new PreferencesTab EntryTab
