+++ src/main/java/org/jabref/gui/preferences/entry/EntryTabViewModel.java
@@ -1,0 +1,83 @@
+package org.jabref.gui.preferences.entry;
+
+import javafx.beans.property.BooleanProperty;
+import javafx.beans.property.SimpleBooleanProperty;
+import javafx.beans.property.SimpleStringProperty;
+import javafx.beans.property.StringProperty;
+
+import org.jabref.gui.preferences.PreferenceTabViewModel;
+import org.jabref.logic.preferences.OwnerPreferences;
+import org.jabref.logic.preferences.TimestampPreferences;
+import org.jabref.preferences.BibEntryPreferences;
+
+public class EntryTabViewModel implements PreferenceTabViewModel {
+
+    private final StringProperty keywordSeparatorProperty = new SimpleStringProperty("");
+    private final BooleanProperty markOwnerProperty = new SimpleBooleanProperty();
+    private final StringProperty markOwnerNameProperty = new SimpleStringProperty("");
+    private final BooleanProperty markOwnerOverwriteProperty = new SimpleBooleanProperty();
+    private final BooleanProperty addCreationDateProperty = new SimpleBooleanProperty();
+    private final BooleanProperty addModificationDateProperty = new SimpleBooleanProperty();
+
+    private final BibEntryPreferences bibEntryPreferences;
+    private final OwnerPreferences ownerPreferences;
+    private final TimestampPreferences timestampPreferences;
+
+    public EntryTabViewModel(BibEntryPreferences bibEntryPreferences, OwnerPreferences ownerPreferences, TimestampPreferences timestampPreferences) {
+        this.bibEntryPreferences = bibEntryPreferences;
+        this.ownerPreferences = ownerPreferences;
+        this.timestampPreferences = timestampPreferences;
+    }
+
+    @Override
+    public void setValues() {
+        keywordSeparatorProperty.setValue(bibEntryPreferences.getKeywordSeparator().toString());
+
+        markOwnerProperty.setValue(ownerPreferences.isUseOwner());
+        markOwnerNameProperty.setValue(ownerPreferences.getDefaultOwner());
+        markOwnerOverwriteProperty.setValue(ownerPreferences.isOverwriteOwner());
+
+        addCreationDateProperty.setValue(timestampPreferences.shouldAddCreationDate());
+        addModificationDateProperty.setValue(timestampPreferences.shouldAddModificationDate());
+    }
+
+    @Override
+    public void storeSettings() {
+        bibEntryPreferences.keywordSeparatorProperty().setValue(keywordSeparatorProperty.getValue().charAt(0));
+
+        ownerPreferences.setUseOwner(markOwnerProperty.getValue());
+        ownerPreferences.setDefaultOwner(markOwnerNameProperty.getValue());
+        ownerPreferences.setOverwriteOwner(markOwnerOverwriteProperty.getValue());
+
+        timestampPreferences.setAddCreationDate(addCreationDateProperty.getValue());
+        timestampPreferences.setAddModificationDate(addModificationDateProperty.getValue());
+    }
+
+    public StringProperty keywordSeparatorProperty() {
+        return keywordSeparatorProperty;
+    }
+
+    // Entry owner
+
+    public BooleanProperty markOwnerProperty() {
+        return this.markOwnerProperty;
+    }
+
+    public StringProperty markOwnerNameProperty() {
+        return this.markOwnerNameProperty;
+    }
+
+    public BooleanProperty markOwnerOverwriteProperty() {
+        return this.markOwnerOverwriteProperty;
+    }
+
+    // Time stamp
+
+    public BooleanProperty addCreationDateProperty() {
+        return addCreationDateProperty;
+    }
+
+    public BooleanProperty addModificationDateProperty() {
+        return addModificationDateProperty;
+    }
+}
