+++ src/main/java/org/jabref/gui/preferences/general/GeneralTab.java
@@ -1,30 +1,20 @@
 package org.jabref.gui.preferences.general;
 
 import javafx.fxml.FXML;
-import javafx.scene.control.Button;
 import javafx.scene.control.CheckBox;
 import javafx.scene.control.ComboBox;
-import javafx.scene.control.TextField;
 
-import org.jabref.gui.Globals;
-import org.jabref.gui.actions.ActionFactory;
-import org.jabref.gui.actions.StandardActions;
-import org.jabref.gui.help.HelpAction;
 import org.jabref.gui.preferences.AbstractPreferenceTabView;
 import org.jabref.gui.preferences.PreferencesTab;
-import org.jabref.gui.util.IconValidationDecorator;
 import org.jabref.gui.util.ViewModelListCellFactory;
-import org.jabref.logic.help.HelpFile;
 import org.jabref.logic.l10n.Language;
 import org.jabref.logic.l10n.Localization;
 import org.jabref.model.database.BibDatabaseMode;
 
 import com.airhacks.afterburner.views.ViewLoader;
-import de.saxsys.mvvmfx.utils.validation.visualization.ControlsFxVisualizer;
 
 public class GeneralTab extends AbstractPreferenceTabView<GeneralTabViewModel> implements PreferencesTab {
 
-    private final ControlsFxVisualizer validationVisualizer = new ControlsFxVisualizer();
     @FXML private ComboBox<Language> language;
     @FXML private ComboBox<BibDatabaseMode> biblatexMode;
     @FXML private CheckBox inspectionWarningDuplicate;
@@ -32,12 +22,6 @@ public class GeneralTab extends AbstractPreferenceTabView<GeneralTabViewModel> i
     @FXML private CheckBox memoryStickMode;
     @FXML private CheckBox collectTelemetry;
     @FXML private CheckBox showAdvancedHints;
-    @FXML private CheckBox markOwner;
-    @FXML private TextField markOwnerName;
-    @FXML private CheckBox markOwnerOverwrite;
-    @FXML private Button markOwnerHelp;
-    @FXML private CheckBox addCreationDate;
-    @FXML private CheckBox addModificationDate;
 
     public GeneralTab() {
         ViewLoader.view(this)
@@ -51,7 +35,7 @@ public class GeneralTab extends AbstractPreferenceTabView<GeneralTabViewModel> i
     }
 
     public void initialize() {
-        this.viewModel = new GeneralTabViewModel(dialogService, preferencesService, preferencesService.getGeneralPreferences(), preferencesService.getTelemetryPreferences(), preferencesService.getOwnerPreferences(), preferencesService.getTimestampPreferences());
+        this.viewModel = new GeneralTabViewModel(dialogService, preferencesService.getGeneralPreferences(), preferencesService.getTelemetryPreferences());
 
         new ViewModelListCellFactory<Language>()
                 .withText(Language::getDisplayName)
@@ -70,19 +54,5 @@ public class GeneralTab extends AbstractPreferenceTabView<GeneralTabViewModel> i
         memoryStickMode.selectedProperty().bindBidirectional(viewModel.memoryStickModeProperty());
         collectTelemetry.selectedProperty().bindBidirectional(viewModel.collectTelemetryProperty());
         showAdvancedHints.selectedProperty().bindBidirectional(viewModel.showAdvancedHintsProperty());
-
-        markOwner.selectedProperty().bindBidirectional(viewModel.markOwnerProperty());
-        markOwnerName.textProperty().bindBidirectional(viewModel.markOwnerNameProperty());
-        markOwnerName.disableProperty().bind(markOwner.selectedProperty().not());
-        markOwnerOverwrite.selectedProperty().bindBidirectional(viewModel.markOwnerOverwriteProperty());
-        markOwnerOverwrite.disableProperty().bind(markOwner.selectedProperty().not());
-
-        addCreationDate.selectedProperty().bindBidirectional(viewModel.addCreationDateProperty());
-        addModificationDate.selectedProperty().bindBidirectional(viewModel.addModificationDateProperty());
-
-        ActionFactory actionFactory = new ActionFactory(Globals.getKeyPrefs());
-        actionFactory.configureIconButton(StandardActions.HELP, new HelpAction(HelpFile.OWNER, dialogService), markOwnerHelp);
-
-        validationVisualizer.setDecoration(new IconValidationDecorator());
     }
 }
