+++ src/main/java/org/jabref/gui/preferences/groups/GroupsTabViewModel.java
@@ -2,10 +2,7 @@ package org.jabref.gui.preferences.groups;
 
 import javafx.beans.property.BooleanProperty;
 import javafx.beans.property.SimpleBooleanProperty;
-import javafx.beans.property.SimpleStringProperty;
-import javafx.beans.property.StringProperty;
 
-import org.jabref.gui.DialogService;
 import org.jabref.gui.groups.GroupViewMode;
 import org.jabref.gui.groups.GroupsPreferences;
 import org.jabref.gui.preferences.PreferenceTabViewModel;
@@ -16,13 +13,10 @@ public class GroupsTabViewModel implements PreferenceTabViewModel {
     private final BooleanProperty groupViewModeUnionProperty = new SimpleBooleanProperty();
     private final BooleanProperty autoAssignGroupProperty = new SimpleBooleanProperty();
     private final BooleanProperty displayGroupCountProperty = new SimpleBooleanProperty();
-    private final StringProperty keywordSeparatorProperty = new SimpleStringProperty("");
 
-    private final DialogService dialogService;
     private final GroupsPreferences groupsPreferences;
 
-    public GroupsTabViewModel(DialogService dialogService, GroupsPreferences groupsPreferences) {
-        this.dialogService = dialogService;
+    public GroupsTabViewModel(GroupsPreferences groupsPreferences) {
         this.groupsPreferences = groupsPreferences;
     }
 
@@ -40,7 +34,6 @@ public class GroupsTabViewModel implements PreferenceTabViewModel {
         }
         autoAssignGroupProperty.setValue(groupsPreferences.shouldAutoAssignGroup());
         displayGroupCountProperty.setValue(groupsPreferences.shouldDisplayGroupCount());
-        keywordSeparatorProperty.setValue(groupsPreferences.getKeywordSeparator().toString());
     }
 
     @Override
@@ -48,7 +41,6 @@ public class GroupsTabViewModel implements PreferenceTabViewModel {
         groupsPreferences.setGroupViewMode(groupViewModeIntersectionProperty.getValue() ? GroupViewMode.INTERSECTION : GroupViewMode.UNION);
         groupsPreferences.setAutoAssignGroup(autoAssignGroupProperty.getValue());
         groupsPreferences.setDisplayGroupCount(displayGroupCountProperty.getValue());
-        groupsPreferences.keywordSeparatorProperty().setValue(keywordSeparatorProperty.getValue().charAt(0));
     }
 
     public BooleanProperty groupViewModeIntersectionProperty() {
@@ -65,9 +57,5 @@ public class GroupsTabViewModel implements PreferenceTabViewModel {
 
     public BooleanProperty displayGroupCount() {
         return displayGroupCountProperty;
-    }
-
-    public StringProperty keywordSeparatorProperty() {
-        return keywordSeparatorProperty;
     }
 }
