+++ src/main/java/org/jabref/preferences/PreferencesService.java
@@ -49,6 +49,8 @@ public interface PreferencesService {
 
     InternalPreferences getInternalPreferences();
 
+    BibEntryPreferences getBibEntryPreferences();
+
     JournalAbbreviationPreferences getJournalAbbreviationPreferences();
 
     void storeKeyBindingRepository(KeyBindingRepository keyBindingRepository);
